objects/bridge: combine bridge modules
This combines each of the bridge type modules into one, given their
only difference is in how the offset height is calculated.
