+++ src/meson.build
@@ -515,10 +515,7 @@ common_sources = [
   'trx/game/objects/general/big_bowl.c',
   'trx/game/objects/general/bird_tweeter.c',
   'trx/game/objects/general/boat.c',
-  'trx/game/objects/general/bridge_common.c',
-  'trx/game/objects/general/bridge_flat.c',
-  'trx/game/objects/general/bridge_tilt1.c',
-  'trx/game/objects/general/bridge_tilt2.c',
+  'trx/game/objects/general/bridge.c',
   'trx/game/objects/general/cabin.c',
   'trx/game/objects/general/camera_target.c',
   'trx/game/objects/general/carcass.c',
